fin proj

+++ ms-tarifas/src/main/java/cl/hilton/tarifas/dto/ProjTipoHabitacionResponse.java
@@ -9,7 +9,7 @@ public class ProjTipoHabitacionResponse {
 
     private String descripcion;
 
-    private Short capacidadMax;
+    private Integer capacidadMax;
 
     private Boolean activo;
 }
